Added lines to read param 'CurrentTrainingStage'
* This gives explicit control over training stage. This replaces the
previous approach of inferring the state implicitly based on the values
of the other params.

* The value of "CurrentTrainingStage" is the index of "trainingstages"
you would like the task to be set to.

* Each time the automator is called, it reads
trials.currentTrainingStage and if necessary, changes the trial.[] to
fit the params of the intended stage.

The flag need2writeParameters is set to 1 if there is a discrepancy.