Added new params, fixation grid overlay, OR automator, and imagedisplay.js
(1) Added two new parameters that give user more explicit control of
display grid:

ngridpoints - a scalar controlling the number of gridpoints in both
directions (square grid; 3x3 grid was the hard coded default used
previously). ngridpoints can be useful if want to have more choice
positions available on the response screen. A 3x3 screen only supports
up to 8AFC.

gridscale - intergridpoint spacing or resolution. If gridscale=1, then
spacing between gridpoints is equal to the width of an image. Images
will be adjacent when gridscale=1 (previous hard coded default).
Gridscale > 1 results in larger grid spacing and a larger grid.
Gridscale < 1 leads to finer grid spacing and may be useful if for
varying the position of the sample.

(2) Added three new parameters that give user more explicit control
over fixation behavior:

fixationScale - The size of the fixation relative to the size of the
sample images. fixationScale=1 means the fixation dot will have the
same diameter as the width of an image. This replaces fixationradius as
the parameter for controlling the size of the fixation.

fixationusessample - now can display the sample image instead of a
fixation dot for user to touch. The sample image will be displayed at
fixationScale x original image width.

fixedratio - number of times the fixation dot needs to be touched to
initiate a trial. More effort engaging the fixation especially when the
sample image is the fixation has been shown to help subjects learn
match to sample tasks.

(3) Added overlay of grid points and bounding boxes for
fixation/sample/test images during test trial phase. This allows the
user to see where elements will render on the screen and make any
necessary adjustments to the grid size or image sizes. Overlay
disappears once actual trials and data collection begin.

(4) Updated automator for training naive subject on 24-object
"objectome" match-to-sample task. The new OR automator replaces
updateTask3 and follows the format of the SR automator. Deleted
updateTask1, updateTask2, and updateTask3.

(5) Created imagedisplay.js and placed canvas setup & rendering
functions there. Modular organization mkturk will aim to provide a
natural breakdown for users learning javascript for mkturk in the
future modular tutorial.

1. Display an image (Canvases) - mkturkimagedisplay.js
2. Wait for user action (Promises & Generators) - mkturk.js
3. Save data (Dropbox) - mkturkdropbox.js
4. Control external hardware (Web Bluetooth) - mkturkble.js
5. Plot data (Googlecharts) - liveplot.html & liveplot_googlecharts.js
6. Load parameters/images externally (Dropbox) - mkturkdropbox.js