Added ImageBuffer, a rudimentary single buffer
Current usage:
* At each trial, an image is requested frmo the ImageBuffer object via
.get_by_name
* On demand, each image is loaded from Dropbox and placed into cache,
if not there already
* If in cache, just return the image  from there.
* In future, if you know what the images needed for the first N trials
needs to be, can use .buffer_these_images(imagename_list) to manually
buffer a particular set of images
* Can call await buffer_chunk(chunksize) to await for chunksize images
to be added to cache. Should be called in downtimes. Not sure if this
will work asynchronously in background without messing up
.get_by_filename, but maybe it does

Object design description from comments:
* This object downloads, caches, and gives out images based on their
filenames.
* Nothing else. No labels, no grouping, nothing. Just images and their
filenames.

* Single buffer for now for finite list of images that can completely
fit in device RAM.
Todo:
* double buffer to serve imagelists greater in size than device RAM
* make this Dropbox independent - make this usable with local disk, or
in-lab server, for example