Updates to new drag 'n drop version of mkturk
This commit modifies naming in ENV, TASK, and TRIAL globals to match
format of previous names. The main reason for this is for new data
files to match previous data files generated by mkturk.

Introduced global variable CURRTRIAL to bind together all global vars
for the current trial. Info in CURRTRIAL feeds TRIAL which is saved to
the datafile.

Name changes are not set in stone quite yet and could be adjusted in
finalizing this new version.

Re-implemented:
- NFixations
- GridScale

De-implemented:
- render photoreward and render eyefixation (not used)
- sampleBlock
- mouse listeners (will re-implement once have touch listeners worked
out)

Other changes:
- ordered operations and added comments so that main loop in mkturk
reads even more logically
- consolidated reward/punish code in mkturk.html and cut some legacy
pieces
- switched to ENV.CanvasRatio instead of ENV.CanvasScale. As the
inverse of CanvasScale, CanvasRatio may be slightly more intuitive
since it tells you the ratio between the virtual screen and the
physical screen. CanvasRatio = 0.5 means that the Canvas is half the
size of the Physical screen in pixels.
- converted to explicit naming and handling of canvases for display
- do not read performance history if TASK.Automator=0