Apply material only to "Base" mesh
Threejs material in object's scenebag params is applied only to the Base mesh. This is for mesh files with multiple part.
The remaining sub-meshes of the object have their material specified during object gltf file creation.
For example, a face has white spheres for the eyeballs and black spheres for the pupil, so mkturk_threejs will not/cannot adjust the material of these components, only the Base head mesh