MAJOR: 3D rendering using three.js (Part 4 of 5) -- Saving images
Saves out rendered sample scene to a png file on cloud storage. Added different trial sampling strategies (uniform_without_replacement & sequential) to default mkturk sampling (uniform_with_replacement). "Sequential" sampling is good for saving out images because it will keep them in the order specified in the scenebag (eg, if monotically varying pose, size, etc) instead of random draws. Sampling strategy is now an explicit parameter in the task params file (TASK.SamplingStrategy).

When Task.Agent is specified as "SaveImages", mkturk will grab two screenshots: one of the webgl canvas render to get the foreground 3d scene and one of the background 2D image canvas. These two png files can be composed to create the image. The whole canvas is saved so that the positions of the objects are preserved. However, this does create large images. Importantly, the javascript canvas export function is limited to 96 dpi which is fairly coarse. To improve on this, we scale the threejs rendering canvas by 4x to get crisper renders both on-screen and in the final screenshot (now the png is 4x bigger). We are capped at 4096 pixel size for images by webgl, so using 5x or more resolution starts becoming unstable. The renderratio is hard-coded (ENV.ThreeJSRenderRatio = 4) but in the future could be user specified when rendering out high-quality images.

*Note, because the output image that is saved reflects the size of the canvas for that display (eg, pixel c tablet has larger canvas than a pixel 4 phone), it is advisable to be consistent about rendering out images from a particular device; ideally, the one used for the experiment and not just any device. The pixel c is a good choice given its high res, large canvas which produces good saved image exports.

INTERTRIAL INTERVAL: added an intertrial interval at the very end of the mkturk.html task code. Will pause for TASK.InterTrialInterval milliseconds if this variable is specified in the params file; otherwise, defaults to not pausing between trials. To be efficient, the pause is shortened depending on how long the housekeeping at the end of a trial took, that way the total time between giving reward/punish to completing (housekeeping + pause) equals the desired intertrial interval.

MINOR
- changes to eyetracking code to help debugging. Will be changed in future commits.
- renderer.gammaoutput replaced by renderer.outputEncoding. Default "Linear" does not produce natural color. Instead used THREE.sRGBEncoding.