Fix 3d canvas bug on task parameter changes
"The problem is that when the parameter changes, canvasvisiblewebgl’s size defaults to the original canvas size, not taking into account ENV.THREEJSrenderratio which is currently 4. If you check the size of canvasvisiblewebgl after you change any param, you can see that it’s not 4*768 but 768. If you make ENV.THREEJSRenderratio=1, the problem disappears."

Solved by putting setupCanvas & initThreeJS at same location so that webgl canvas size is not reset by setupCanvas unless followed by a setRenderRatio in initThreeJS (ie, a completely new scene is being loaded because the sceneparams sample or test bag was changed)

Co-Authored-By: younahjeon <younahjeon@users.noreply.github.com>