Add fixation dot, add reaction time histogram to liveplot
Add fixtion dot overlay. New parameter FixationDotSizeInches.

Add fixation window. New parameter FixationWindowSizeInches (SampleFixationSizeInches is deprecated).

If FixationWindowSizeInches>0, this is used for fixation & sample screen bounding boxes overriding the fixation or sample display element's bounding box.

Both fixation dot & fixation window are applied to the fixation and sample screens (not to test or choice screens) and are intended for fixation tasks such as RSVP.

Cleaned up the way mkturk saves timestamps in a fixation task and how it specifies Response & Correctitem so that it's easier for liveplot & fireplace to get performance stats.

Added a basic histogram to report the time of the touch from the display of the element onscreen. This is particularly useful for fixation tasks but also gives reaction time on choice tasks.

Turned off yellow dot plotting of eye trace during data saving.